OLD: Minor updated EC_GUI_MATLAB_V1.zip file (brief "quick-start" guide included in .zip)
NEW: updated gantry/survey value files based on Hamamatsu Electrical R0, R1-R4 SFU sensors, and Hybrid R0/R5 Mechanical Sensors. **There are 2 "gantryValues" files --> the R4 powerboard may interfere with the pickup tool, "gantryValues2" has a modified tool location on that particular sensor to accommodate this if need be (test with a dry placement to see if it is necessary to use).